requirements.txt
numpy==1.26.3
PyYAML==6.0.1
